orbit_ai_backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.0.0,<3.0.0
httpx>=0.26.0,<1.0.0
groq>=0.4.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
